updated gitignore to add the products of slatex.   .gitignore
changed the user guide to use the more rackety slatex, instead of stex.
  doc/Makefile, doc/user-guide.pdf, doc/user-guide.tex

removed gentemp in favor of racket's built-in generate-temporary.
  nanopass/helpers.rkt, nanopass/pass.rkt, nanopass/meta-parse.rkt

paired down the contents of records to remove identifiers that are not needed
across uses of the language record to make it easier to reconstruct the record.
also moved functions from language.rkt into records.rkt to more centrally
locate them.
  nanopass.rkt, nanopass/language-node-counter.rkt, nanopass/language.rkt,
  nanopass/meta-parser.rkt, nanopass/parser.rkt, nanopass/pass.rkt,
  nanopass/records.rkt, nanopass/unparser.rkt